# api/openapi.yaml
openapi: 3.0.0
info:
  title: PC Shop API
  version: 0.0.0
...
paths:
...
  /api/orders:
    post:
      operationId: Orders_create
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
    get:
      operationId: Orders_list
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/orders/{id}:
    get:
      operationId: Orders_get
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '401':
          description: Access is unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: {}
      description: Единый формат ошибки для всех ресурсов API.
...
    CreateOrderRequest:
      type: object
      required:
        - items
        - method
        - recipientName
        - phone
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemInput'
        method:
          type: string
          enum:
            - delivery
            - pickup
        recipientName:
          type: string
        phone:
          type: string
        address:
          type: string
...
    Money:
      type: integer
      format: int32
      minimum: 0
      description: Целая сумма в рублях, без копеек — в магазине одна валюта.
    Order:
      type: object
      required:
        - id
        - method
        - recipientName
        - phone
        - status
        - total
        - createdAt
        - items
      properties:
        id:
          type: integer
          format: int32
        method:
          type: string
          enum:
            - delivery
            - pickup
        recipientName:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum:
            - paid
        total:
          $ref: '#/components/schemas/Money'
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      required:
        - name
        - price
        - qty
      properties:
        productId:
          type: integer
          format: int32
        name:
          type: string
        price:
          $ref: '#/components/schemas/Money'
        qty:
          type: integer
          format: int32
          minimum: 1
...
    OrderItemInput:
      type: object
      required:
        - productId
        - qty
      properties:
        productId:
          type: integer
          format: int32
        qty:
          type: integer
          format: int32
          minimum: 1
...